requirements.txt
Pillow
numpy
pyautogui
opencv-python